begin rewrite in rust
Why? Because I want to add features that need memory
allocations. But C has `malloc`! Yes, but we can't
use C for this, because when we link in `libc`, it
segfaults somewhere in `__ctype_init` early in the
initialization (before main is even ran)! I'd really
like to write some C again, but... :|